+++ src/main/java/uk/co/shockwaveInteractive/ShockMetalMain.java
@@ -5,7 +5,6 @@ import net.minecraft.item.ItemGroup;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.InterModComms;
 import net.minecraftforge.fml.client.registry.RenderingRegistry;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
@@ -16,17 +15,18 @@ import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import net.minecraftforge.fml.common.Mod;
-import uk.co.shockwaveinteractive.init.Entities;
+import uk.co.shockwaveinteractive.config.MainConfig;
 import uk.co.shockwaveinteractive.tabs.ShockMetalTab;
 import uk.co.shockwaveinteractive.util.reference.MainReference;
 import uk.co.shockwaveinteractive.util.handlers.RegistryHandler;
 import uk.co.shockwaveinteractive.util.renderers.SpriteRendererShock;
+import uk.co.shockwaveinteractive.util.renderers.VacuumMinecartRenderer;
 import uk.co.shockwaveinteractive.world.gen.WorldGenCustomOres;
 
 import java.util.Random;
-import java.util.stream.Collectors;
 
 import static uk.co.shockwaveinteractive.init.Entities.SHOCK_GRENADE_ENTITY;
+import static uk.co.shockwaveinteractive.init.Entities.VACUUM_MINECART_ENTITY;
 
 @Mod(MainReference.MODID)
 public class ShockMetalMain
@@ -46,6 +46,8 @@ public class ShockMetalMain
 		FMLJavaModLoadingContext.get().getModEventBus().addListener(this::processIMC);
 		// Register the doClientStuff method for modloading
 		FMLJavaModLoadingContext.get().getModEventBus().addListener(this::doClientStuff);
+
+		MainConfig.register();
 
 		RegistryHandler.init();
 
@@ -77,15 +79,15 @@ public class ShockMetalMain
 	private void enqueueIMC(final InterModEnqueueEvent event)
 	{
 		// some example code to dispatch IMC to another mod
-		InterModComms.sendTo("examplemod", "helloworld", () -> { LOGGER.info("Hello world from the MDK"); return "Hello world";});
+//		InterModComms.sendTo("examplemod", "helloworld", () -> { LOGGER.info("Hello world from the MDK"); return "Hello world";});
 	}
 
 	private void processIMC(final InterModProcessEvent event)
 	{
 		// some example code to receive and process InterModComms from other mods
-		LOGGER.info("Got IMC {}", event.getIMCStream().
-				map(m->m.getMessageSupplier().get()).
-				collect(Collectors.toList()));
+//		LOGGER.info("Got IMC {}", event.getIMCStream().
+//				map(m->m.getMessageSupplier().get()).
+//				collect(Collectors.toList()));
 	}
 	// You can use SubscribeEvent and let the Event Bus discover methods to call
 	@SubscribeEvent
@@ -103,5 +105,6 @@ public class ShockMetalMain
 
 	private void registerEntityRenderingHandlers() {
 		RenderingRegistry.registerEntityRenderingHandler(SHOCK_GRENADE_ENTITY.get(), SpriteRendererShock::new);
+		RenderingRegistry.registerEntityRenderingHandler(VACUUM_MINECART_ENTITY.get(), VacuumMinecartRenderer::new);
 	}
 }