+++ src/main/java/uk/co/shockwaveInteractive/objects/blocks/item/BlockItemBase.java
@@ -9,6 +9,6 @@ public class BlockItemBase extends BlockItem
 {
 
 	public BlockItemBase(Block block) {
-		super(block, new Item.Properties().group(ShockMetalMain.SHOCKMETALTAB));
+		super(block, new Item.Properties().group(ShockMetalMain.SHOCKMETALTAB).isImmuneToFire());
 	}
 }