+++ src/main/java/uk/co/shockwaveInteractive/util/reference/GeneralReference.java
@@ -1,8 +1,0 @@
-package uk.co.shockwaveinteractive.util.reference;
-
-import net.minecraftforge.registries.ObjectHolder;
-
-
-@ObjectHolder(MainReference.MODID)
-public class GeneralReference {
-}